Add state to distributed composable API
[ghstack-poisoned]
